+++ Programs/DviWare/dvipdfm-x/source/pdfdoc.h
@@ -1,6 +1,6 @@
 /* This is dvipdfmx, an eXtended version of dvipdfm by Mark A. Wicks.
 
-    Copyright (C) 2007-2016 by Jin-Hwan Cho and Shunsaku Hirata,
+    Copyright (C) 2007-2017 by Jin-Hwan Cho and Shunsaku Hirata,
     the dvipdfmx project team.
     
     Copyright (C) 1998, 1999 by Mark A. Wicks <mwicks@kettering.edu>
@@ -58,7 +58,7 @@ extern pdf_obj *pdf_doc_get_reference  (const char *category);
 
 extern int      pdf_doc_get_page_count (pdf_file *pf);
 extern pdf_obj *pdf_doc_get_page (pdf_file *pf, int page_no, int options,
-				  pdf_rect *bbox, pdf_obj **resources_p);
+				  pdf_rect *bbox, pdf_tmatrix *matrix, pdf_obj **resources_p);
 
 extern int      pdf_doc_current_page_number    (void);
 extern pdf_obj *pdf_doc_current_page_resources (void);