+++ Programs/DviWare/dvipdfm-x/source/spc_util.c
@@ -42,7 +42,6 @@
 #include <miktex/unxemu.h>
 #endif
 
-
 #ifndef ISBLANK
 #define ISBLANK(c) ((c) == ' ' || (c) == '\t')
 #endif