+++ Programs/TeXAndFriends/xetex/source/XeTeXFontInst.cpp
@@ -42,7 +42,6 @@ authorization from the copyright holders.
 #if defined(MIKTEX)
 #include "miktex-first.h"
 #endif
-
 #include <w2c/config.h>
 #include <kpathsea/kpathsea.h>