+++ Programs/TeXAndFriends/xetex/source/XeTeXFontMgr_FC.cpp
@@ -33,7 +33,6 @@ authorization from the copyright holders.
 #if defined(MIKTEX)
 #include "miktex-first.h"
 #endif
-
 #include <w2c/config.h>
 
 #include "XeTeXFontMgr_FC.h"