+++ Programs/TeXAndFriends/xetex/source/XeTeX_ext.c
@@ -39,7 +39,6 @@ authorization from the copyright holders.
 #if defined(MIKTEX)
 #include "miktex-first.h"
 #endif
-
 #include <w2c/config.h>
 
 #include <poppler-config.h>